Improved content script matches.
When your session in TUWEL expires and you try to
access the page it will redirect you to the login page
with additional url parameters ("?errorcode=4"). Up until
now this page didn't get matched and the content scripts
did not execute.

This is fixed now and content scripts will
run on TUWEL login pages with error parameters in the
url as well.